katcp2spead bridge: transmitter host and port set in session.py
Originally the spead_host and spead_port were set through the katcp2spead.py
script, when the device is added in base_cam_processes.conf
The later requirement was that multiple transmitters could exist for
the sensor stream, and that they could be started and stopped independently.
Consequently, the spead_host and spead_port are now no longer set when the
device server is started. They are set when the transmitters are added
in session.py.
In this fix the spead_host and spead_port variables are removed from the
Katcp2SpeadDeviceServer init. They are left in the parser options
in the katcp2spead.py script for the moment, for (broken) backwards
compatbility.