Allow use of fake CAM sensor server in katcp2spead
To make testing and simulation easier, katcp2spead can now receive
sensor updates from a single device faking the entire CAM system.
This is simplified by sending the sensor group to the underlying
DeviceServer instead of a KAT connection object, as this can handle
both the real and fake scenarios.