Variable renames and doc updates
Replace SPEAD "transmitter" with "destination" as this seems to make
more sense at the KATCP command level (the transmitter is an internal implementation feature that the user does not need to know about).
Replace "tx_period" with "flush_period" to be more descriptive and
match usage in other places. Use colon to separate IP and port in
destination name.