New sim_observe script
This script drives the simulation of the CBF and CAM subsystems as the
main inputs into the ingest subsystem. This allows for standalone
testing of the ingestor. The sim_observe script starts a device server
(FakeCamSensorServer) that provides all the sensors that the real CAM
system would provide to katcp2spead. These sensors are then updated
from a text file containing the sensor values of a real dataset, in
effect replaying the observation. The script also starts capturing on
the CBF, ingest and katcp2spead in the form of a mini capture session.

The provided sensor event file (rts_scp_events.txt.gz) does a simple
track of the SCP for 40 seconds, using the two proposed antennas of
the RTS system.