Moved to HDF5 v3.0 and included many bug fixes after first simulator integration
Lots of little fixes, but the major change is a complete overhaul of /MetaData.
We know write the telescope model in /TelescopeModel with each component
having it attributes and datasets stored herein.

Everything is now either an attribute or a dataset (sensor) and we expect
to get these pieces of information via spead from either CBF or TM.

Still no doubt lots of fixes to come