Add support for CAM attributes
The new TelescopeModel in katsdpingest distinguishes between static
data items (*attributes*) that remain constant for the duration of
an experimental run and dynamic data items (*sensors*) that change
as a function of time. While the latter is realised by the standard
KATCP sensors provided by CAM, the former is more nebulous and akin
to the CAM configuration.

This adds support for attributes. These are extracted from the CAM
configuration (currently only antenna descriptions) or faked along
with sensor values as a key-value dict at the top of the events file.
The FakeCamEventServer provides these attributes upon request.
The event file parser is spun out and improved to support attributes
and comments. Updated example events file to contain attributes.

The katcp2spead server is renamed to cam2spead to reflect its expanded
role. It inserts the received attributes in the initial SPEAD heap
since their values will never change.