Send SPEAD stop packet to close CAM stream
When the SPEAD transmitter thread finishes, send the SPEAD stop packet
to the CAM stream destination so that the receiver can shut down
gracefully. Previously this was unnecessary because the CAM and CBF
streams went to the same port (i.e. was the same stream) and the CBF
part of the stream did the stop packet.