Parse sensor value to appropriate type
The sensor listener framework of KATSensor actually returns values as
strings and not the underlying type as previously thought. Use the
proper katcp.Sensor mechanism to parse this string to the correct type
before sending repr(value) as a string to the SPEAD stream, as this
ensures that the type is correctly inferred by eval() on the receiving
end. Improve docstrings to avoid further misguided assumptions.