Make add_destination independent of start_stream
The add_destination() request of cam2spead can now occur before or
after start_stream(), and stop_stream() also removes destinations
automatically. This is more like the behaviour of the CBF with the
usual ordering of capture_destination, capture_start, capture_stop.

To achieve this, the SPEAD transmit thread no longer stores the
initial heap but receives it via the mailbox like all other heaps.
The thread is created when the stream starts instead of when the
destination is added, which allows the device server to handle
multiple stream starts and stops. All threads are properly joined
to ensure clean termination. The timeout on the transmit thread
mailbox is drastically reduced to avoid lengthy delays on
termination.

Renamed periodic_transmit to collate_and_transmit as this is more
descriptive of the purpose of the method and the thread that runs it.