Align stream management interface with CAM
This introduces named SPEAD streams, where the names will be data
product IDs. The internals of the server is still the same, with
all sensors being sent to all destinations. The server only really
works well for one stream currently. The names of the stream
creation, starting and stopping requests are aligned with the CBF
and ingest style (noun-verb).