Remove data product ID from spead item names
The 'label' and 'obs_params' items are supposed to be associated with
a specific capturing session (which currently maps to a sub-array).
This allows it to be reconstructed after the fact from the sensor
store (i.e. during "re-augmentation"). Since we only have a single
sub-array active at any one time for RTS, this naming scheme is
currently not worth the effort.