Added multicast support to ingest
This is done in a bit of a roundabout fashion, but I did not want
to change any of the SPEAD rx code.

Basically we check for the specified data_host address to have a
prefix in a list of known multicast prefixes. If so a stub
socket is created which is used to subscribe to the desired
addresses on the port on which the SPEAD rx will listen.

Format support for ip<+[count]> is provided where the count
specifier will subscribe to count address starting from the
unmolested base address and incrementing the trailing octet
by one up to the count limit.

e.g. '224.1.1.1+2' will suscribe to 224.1.1.1, 224.1.1.2 and 224.1.1.3