Fix spead metadata packet issue
In tandem with a bug introduced in PySPEAD in Feb 2014 (commit 6557167)
this causes the CBF ingest thread to crash upon receiving the labelling
metadata packet. Instead of starting from an empty ItemGroup, this code
reused the data_ig, which already contained the 'timestamp' and 'xeng_raw'
items. While these have already been sent in the first metadata packet,
the PySPEAD bug caused their descriptors to be resent, thereby resetting
the 'timestamp' value to None which then crashes upon ig['timestamp']
access in ingest_threads.py. The solution (apart from fixing PySPEAD)
is to start from a fresh ItemGroup to avoid this.