Refactor KATCP->SPEAD sensor update
Introduce the update_sensor method on Cam2SpeadDeviceServer that can push
KATCP-like sensor updates to a SPEAD stream for both real and virtual KATCP
sensors. By having a central method we can re-use the same ig update code
which was already copied thrice. This is in preparation of adding yet
another virtual sensor (obs_script_log). The debug logging is also improved
to highlight what happened during the sensor update (and virtual sensor
updates get it now too). Half the added lines are in fact docstring lines
for the update_sensor method, so there is a net removal of code.