Specialise SensorBridge for KATCP and virtual ones
The original SensorBridge only dealt with real KATCP sensors. We also want
to handle virtual sensors in the same way, so we change SensorBridge to be
a base class, with KatcpSensorBridge and VirtualSensorBridge handling the
special cases.

In the process, make description an attribute of SensorBridge so that virtual
sensors can have one too (in fact, that's all they have...). The strategy and
param attributes are now set directly instead of via store_strategy.