Register virtual sensors same as KATCP ones
Create VirtualSensorBridge objects for virtual sensors and register them
in the same way as the real KATCP ones. This removes the current hardcoding
of virtual sensors like obs_label, obs_params and obs_script_log, which got
tricky once you add component-dependent sensors like m062_noise_source.

The virtual sensors are now specified in the sensor list / CSV file and all
sensors that are not found on the host object are assumed to be virtual.
This concentrates all the system-specific stuff in the config files and
leaves cam2spead more generic.