Add set_nd_sensor request to cam2spead for RTS
This KATCP request manually sets or clears the noise diode flag for the given
receptor, thereby driving the virtual noise diode sensor. It is typically
called in the RTS observation script during manual noise diode firings to
compensate for the lack of a real noise diode flag from the digitiser.