Rename CBF component to match RTS system
The simplest way to get going with RTS is to rename the CBF component
from 'cbf' to 'data_rts' in order to match the CAM client name. This
ensures that cam2spead delivers the appropriate sensors. The CBF component
is now looked up by class and not name in katdal, so the name is irrelevant.
Also cull KAT-7 specific sensors from the sensor config file.

This approach is OK for now since the rest of the telescope model (e.g.
antennas and enviro) is also hard-coded for RTS anyway and needs fixing too.