Obtain center_freq from CBF
In KAT-7 the center frequency used to be a function of the mixer LO setting
in RFE7. That meant that the ingest needed to be told what it is via KATCP
(typically from the CaptureSession). In RTS this is no longer the case as
there is no mixer. The center frequency is determined by the CBF product and
passed along in the SPEAD metadata to the ingest, becoming a required sensor.

This mainly fixes the frequency axis in the signal displays (and removes
some cruft).

(Reviewed by sratcliffe)