Fix autotune_mkimage.py to not leave a volume behind
Specifying a volume when creating the container leaves it as a mount
point in the resulting image, which then causes problems when trying to
launch the image. Specifying only the bind seems to work, while being
transient and not captured in the committed image.

Also some minor updates to avoid deprecated functionality.