Make SD handling more like master was
It was proving tricky to replicate the exact semantics of the metadata
reissues with the modified design. Rather than go down the rabbit-hole
of figuring out what can safely be changed, the design has been changed
to reflect the old one i.e. every request for metadata reissue will
recreate the item group (and now heap generator). This needs spead2
0.2.0, so that heap cnts are not reused.