Fix thread safety issue with update_sd_metadata
Each time the SD metadata is resent we update the ItemGroup to reflect
any changes in the data shape or structure. This was done by clearing
the existing ItemGroup and then recreating it.

This was not thread safe since the katcp thread can call update_sd_metadata
and the main thread needs to write entries to it.