Switch from description attr to observer sensor
The antenna description attributes are based on the katconfig values but are
only loaded upon cam2spead startup and then never updated again. This is an
issue if e.g. the pointing model is updated in the proxy but cam2spead is not
restarted, resulting in stale antenna descriptions in the HDF5 file.

The observer sensor, on the other hand, is always the live antenna description
string in use in the antenna proxy and therefore more critical, so use that
instead. This also bumps the file format version to 3.1 (mild sensor change).

This resolves JIRA ticket MRTS-232.