Add basic tilt sensors
These sensors are useful to know when fitting a pointing model to the data.
There might be more tilt sensors to store in future (param-an0 and param-aw0
springs to mind, but these are not currently implemented in the RTS AP).

The file format version can remain at 3.1 as this commit is part of the same PR.

This addresses JIRA ticket MRTS-262.