Register missing sensors for file version 3.1
This adds the tilt and Ku frequency sensors to the telescope model so that
they actually make it into the file. The latter required a new telescope
component for the ancillary proxy ('anc') since Enviro only exposes the
lower-level 'anc_asc' (which seemed like a good idea at the time).