Ingest health sensor
device-status is set to degraded if the CAMIngest thread throws an
exception, and to fail if the CBFIngest thread throws an exception.

At the moment the sensor will never come back to okay, even if
one stops the capture and starts a new one. My assumption is that if it
broke the first time, it is highly likely to break the next time as
well, and hence the unit is unusable.