Make interaction with katsdpsigproc thread-safe
This added a lock in the CBFIngestThread, which is taken for operations
on the shared state.

A lot of the initialisation code moves from _run to __init__, so that it
is guaranteed to be initialised before the next katcp request comes in.

The change to use spead2's multicast support (#87) is also folded up in
here, because it would have been more difficult to do without it, and
it's going to have a giant merge conflict otherwise.

There is also a bug fix: previously drop_sdisp_ip would only drop an IP
from the current capture session, but it would come back from the dead
on the next capture.