Run processing asynchronously to heap fetching
At the moment this doesn't really do anything useful except make for
deeper buffers, because there is only one resource and it is used
more-or-less synchronously. However, this change paves the way for
further optimisations.

The flush handlers are also not yet fully converted to the written
design, since they do not spawn separate output tasks.