Compress output weights
For compression, the actual weight is represented as the product of a
per-channel float32 and a per-channel, per-baseline uint8. This allows
for a wide dynamic range across channels, and a much smaller dynamic
range (8 bits) across the baselines within each channel. The per-channel
value is computed as the maximum weights across all baselines divided by
255, to ensure that the full range of 8-bit values is used.

At present the weights are still set to 1.0 on input, rather than being
estimated from the power in the autocorrelations.

Weights are now also included in the L0 output stream, but not yet in
the signal display output stream.