A support for predefined per-channel flags
These flags are merged with the flags from RFI detection. The code is
also modified to change how multiple flags are combined during
averaging: previously the flags for all visibilities in the group
were ANDed together. Now they are instead ORed together, but in the case
that at least one sample is unflagged, the output flag is zero. With
this change, flags now represent the set of all conditions that
contaminate any of the visibilities used to generate the output
visibility (in the case of at least one unflagged visibility, only the
unflagged visibilities are used, and so flag==0 is correct).

Since the accumulators are now all zero-initialised, the Zero class is
simplified to use the zero method instead of a Fill operation.

At present the per-channel flags are simply set to zero. The results
should be identical to before this change. Future commits will populate
channel_flags with more relevant data.