bf_ingest: fix missed disk checks on dropped heaps
Every 1000 dumps there is a check that there is still free space, and a
log message to indicate dropped heaps. Previously this was done by
checking if the timestamp was on a multiple of 1000 heaps from the
start, but if that particular heap is missed then the check has to way
another 1000 heaps. Fixed it to so that the check will be done on the
next received heap instead.