Remove many Boost dependencies
Boost.Python is replaced by pybind11. Several other dependencies on
Boost are replace by C++11 standard equivalents.

It is not possible to eliminate Boost as a dependency entirely, since it
is used heavily by spead2.

The code gets a bit bigger as a result, because it was necessary to copy
and pybind11-ise the log_function_python class out of spead2. If/when
spead2 switches to pybind11, this can be cleaned up.