Rework the handling of host buffers
This adds buffers for signal display inputs and outputs. It also does
some renaming for consistency, adds a helper _make_resources method, and
fixes a race condition in _frame_job in which the host buffer could be
filled in which the previous values were still being copied to the
device.