cam2telstate no longer waits for subarray to be ready
CAM now guarantee to us that CBF sensors will be ready by the time
data-product-configure is called. We thus don't wait for the subarray to
be active, so that katsdpcontroller can then wait for us to be ready
before starting other services.

Also adds a basic katcp device server with no commands but a sensor to
indicate the status. This will allow it to be harvested by CAM, to help
us debug timing/ordering bugs. It will also be useful for
katsdpcontroller to wait on.