Start katcp server only after initialisation
This will simplify the interaction with katsdpcontroller, which now only
needs to wait until it can connect to the katcp port (which it does for
all katcp-based services anyway). In practice the intermediate states
might not have been observable anyway, because mc only periodically
polls the katcp port and might not have been in time to observe them.