Add pointing and delay adjustment sensors
These reflect the recently added ability to adjust pointing and delays.
In addition, the CBF delay centre frequency was missing (although
the loaded delay correction will also provide that in future).

Also proactively add the imminent receptor data-suspect sensor as
well as the RTS Ku-band signal generator frequency setting. Both will
probably fail with warnings until CAM rolls out the former (and
until installed on RTS for the latter). The latter sensor will
eventually be subsumed in the final CAM-provided centre frequency
sensor, but leave it in as reminder for now (and it fixes katdal too).