Move IngestDeviceServer to its own module
This is to support unit testing it. It's mostly a cut-and-paste job, but
the internal-log-level request was rewritten to use some internals from
the logging module rather than a custom logger class. The custom logger
class was problematic because it would miss loggers that were registered
prior to the logging class being set.