Add --servers and --server-id options
Each ingest process is now expected to get the same options, except for
--server-id, and internally compute its own output ranges and
destination multicast groups. This is similar to how it already
functioned for CBF inputs, but now applies to outputs too.

The metadata being placed in telstate now correctly describes the joint
product rather than the per-server information.