Eliminate --subarray-numeric-id parameter from cam2telstate
katportal can now tell us what it is. It also tells us the names of the
subarray, CBF and SDP proxies without us having to scrape it from
pool_resources.

This is necessary for the future SDP MC API, where the product name is
arbitrary and might not contain the subarray number.