Make spectral and continuum outputs both optional
Instead of defaulting the SPEAD output to localhost, if no SPEAD
endpoint is given, the output is suppressed (and the buffers for
continuum are also suppressed if continuum is disabled).

Suppressing the spectral output could in theory be slightly more
efficient, because the weight compression step could be skipped, but
since that's not expected to be a common use case I haven't tried to
further optimise it (which would uglify the code).