Support capturing a subset of beamformer partitions
Since CBF no longer support ?passband, this implements the functionality
in bf_ingest.py. At the moment it can only be selected at startup time
as a channel range, but it would be fairly easy to add as a katcp
request.

The start channel is stored as an HDF5 attribute in the file. I also
fixed the code that stores the stream name as an attribute: I'm not sure
how that hasn't always caused crashes.

There is also a little bit of flake8 cleanup.