Support getting CBF sensors from fully qualified names
This integrates with recent cam2telstate and katcbfsim changes to
produce these fully-qualified names, in addition to the unqualified
names.

There is still backwards-compatibility support for unqualified names
(they are consulted last), but it will be removed as soon as this has
been proven on site.

Also adds a _src_streams telstate key to indicate which input stream
produced the output stream, to allow e.g. katdal to trace backwards from
a file to determine any properties it needs that aren't available in the
L0 telstate keys.