Use full stream names in bf_ingest
Make bf_ingest use the full name for the antenna-channelised-voltage
stream and its instrument, as determined by the src_streams and
instrument_dev_name keys. This moves the code that was previously added
to ingest_session.py into utils.py so that it can be shared between
ingest and bf_ingest.