Cleaner handling of stream termination
1. Stop the receiver only once we've received a stop item per endpoint,
rather than on the first stop item. This won't make much difference to
real CBF (where all the stop items are sent by the CMC), but will avoid
losing data from the simulator if the split simulators are a bit out of
sync.

2. On capture-done, wait up to 5 seconds for the receiver to shut down
in response to stop heaps before stopping it ourselves. This avoids
losing out on data that is still in flight at the time capture-done is
received.