Report bad heaps via sensors
This makes ingest a lot quieter in the face of bad data. No log messages
are printed if there are incomplete heaps, if the descriptors fail to
arrive etc (they are still logged at debug level). It will still log
if it receives a partial frame e.g. if some X engines are working but
others are not, but this will be at most one message per frame rather
than one per heap.