Provide sensors to track amount of flagging
The signal display flag counter is extended to also count the number of
visibilities per baseline with any flag bit set (in addition to the
number with each flag bit set). This is not currently wired into the
signal display SPEAD stream, but it could be. It's then summed (on the
host) to produce a single number for the total number of flagged
visibilities (output-flagged-total). Another sensor counts the
corresponding number of all visibilities (output-vis-total) so that
ratios can easily be computed for dashboards.