Handle missing and out-of-order heaps in digitiser capture
Previously each packet was expected to immediately follow the previous
one in timestamp. Packets are now allowed to appear in any order (except
that they cannot predate the sync point), and are written at the
appropriate place (possibly creating holes in the file).

At present there is no way to know which heaps went missing, other than
to look for 4096 zeros in the file.