Fix some issues with sensor types
In Python 3, numpy.int64 is not a subclass of int, so cannot currently
be used as a sensor value because aiokatcp doesn't know how to encode
it. This eventually needs to be fixed in aiokatcp to use ABCs.

The increment helper function was also passing the timestamp as a
status.