Fix autotuning script to run the new code
The previous rewrite of the autotune_mkimage had a serious flow in how
it handled --copy: it would run the autotuning for the image from which
the copy was being made, instead of for the target image.

This is now fixed by extracting a tarball from the copy source (by
running a no-op container), and then injecting it into a new container
created from the base image (which has the new code) before autotuning.