Upgrade to latest version of spead2
This is just to avoid bit-rot. There are no new features being used.

Also switch to using the pybind11 vendored within spead2 (which is
vendored here). This avoids the need to have pybind11 installed in the
system.

Some changes were needed to the way logging is handled because it has
been changed in spead2.