Reduce spurious warnings and errors from cam2telstate
- Downgrade the warning when a sensor is a state other than
  nominal/warn/error.
- Remove the warning when a template doesn't expand to anything, which
  would happen for tied-array-channelised-voltage sensors when those
  streams weren't included in the config. It was there as protection
  against mistyping a name in a substitution, which wouldn't error out
  because substitutions were kept in a defaultdict. Change that to a
  regular dict, pre-populated with the known stream types.
- Query source_indices on tied-array-channelised-voltage streams to get
  a list of input indices for which weights exist, instead of trying to
  subscribe to all of them.

The code for querying sensors that are needed to get the list of sensors
to subscribe to has been refactored. Adding a new bootstrap sensor
required changing code in multiple places, but it's now all wrapped up
into a get_sensor_value function, and values aren't carried around in
the object. get_sensors is now a coroutine so that it can query these
values.