Switch to HDF5 1.10 for beamformer capture
This fixes the issues with writing very large attributes to the
/TelescopeState group.

Some code changes were needed because the under-the-hood mechanism I was
using to do direct chunk writes has been replaced in HDF5 1.10.3 with a
H5Dwrite_chunk function, but it still doesn't have a C++ wrapper. We
instead directly call the C function.