Use timestamps from incomplete heaps
To mitigate SR-1440, process timestamps extracted even from incomplete
heaps (which will be present because CBF inlines them in *every* packet).
When things go wrong in the network and all the heaps are incomplete,
this allows a still-buffered frame to get flushed timeously, rather than
waiting for the next complete heap.

This involved some fairly serious refactoring of the receive code. The
upside is that it is now much easier to validate that heaps are not
double-counted (other than input-missing-heaps-total, which is unrelated
to the heaps received).