Catch and report ValueError when updating items
This hasn't actually happened for a very long time, but in theory CBF
could send us data that doesn't match the descriptor and cause a
ValueError in heap processing.

Add a new 'bad-heap' rejection type, and also apply to heaps that have
timestamps but are missing xeng_raw or frequency.