Improvements to counters
The counters are now computed in the receiver thread rather than in the
session thread. This makes it possible to count more events, such as too
old, invalid timestamp, and incomplete heaps.

The downside is that it is probably too expensive to lock the counters
on every heap. The receiver thus keeps a thread-local copy of the
counters which it updates without locks, and periodically schedules a
copy into a public, thread-safe version.

This does mean that there are up to three different clocks updating
counters:
1. The C++ code updating every 1ms.
2. The Python code reading the counters from C++ every 1s.
3. Prometheus scraping every 4s.

That could lead to strange sampling issues.