Encode unicode to UTF-8 in telescope model
bfingest still has legacy telescope model code hanging around. It was
raising exceptions when storing sensors if the values were unicode
strings (and also because the status was Python 3 str). Add a shim to
pass fields through `_array_encode` to encode them to UTF-8.

This isn't perfect because Python 3 code won't decode it again on load
(it'll remain in numpy 'S' type). But since the telescope model is
legacy it's probably not worth mucking about trying to get the right
h5py custom data type incantations.