Prepare bfingest receiver for multi-heap slices
The receiver code is now written for slices that span multiple heaps in
time, but it is left at one heap per slice in config because the rest of
the code is not yet configured for it. This depends on a
not-yet-released version of spead2, so the submodule has been updated to
the development branch.

To clean up the plethora of x_per_y variables, a unit system has been
introduced which also ensures unit correctness at compile time. This adds
more code but the code in receiver.cpp should hopefully be easier to
read. A number of variables will end up as int64 as a result, but we're
on x86_64 so that shouldn't hurt performance.

The unit system is currently isolated within the receiver class, but
could be expanded more generally as needed.