bfingest: Make stats work with multi-heap slices
This is the last component needed to make unit tests work with
heaps_per_slice_time != 1.

The captured_timestamps dataset was removed from the file rather than
trying to fix it, and the version bumped to 4.

The slice class now has a bytes helper that replaces the helpers in
receiver and hdf5_bf_raw_writer.