Decode strings received from katportalclient
With the switch of telstate to msgpack, bytes vs strings gets preserved
more exactly, which leads to things like ingest getting confused by
finding bytes when it expects strings (all of which is thanks to
katportalclient being Py2-only).

Recursively decode all byte strings as UTF-8. There is an option to
override the decoding on a specific sensor if we ever need to store
binary blobs. Numpy arrays do not get converted.