Fix intermittent bfingest test failures
The failures were occurring because while the inproc spead2 transport is
lossless, stopping the stream before all the packets have been processed
causes the remaining ones to be dropped on the floor. There was a sleep
to allow the data to make it, but it was unreliable.

Fixing this required a fair bit of extra plumbing: the spead2-level heap
counter is plumbed into the counter system so that it can be accessed
from the test. The higher-level count of payload heaps is not useful
for this because it doesn't count heaps that are still in the receive
window. Some extra utility code was also added to figure out how many
heaps are expected to arrive, given that the receiver is only
subscribing to a sub-band.