Take per-visibility flags as input
This replaces the per-channel flags, and allows for things like channel
masks that vary depending on baseline length (although that is not
currently implemented). A separate baseline mask is still provided:
it is *not* fed into the flagger, because it may still be useful to
provide RFI flagging for a baseline which is suspected to have a faulty
input.

An extra operation (MergeFlags) is added after the flagger to combine
the input flags, baseline flags and RFI flags into a single flag array.
This simplifies the operation of CountFlags and Accum, which no longer
need to do that merging themselves.