Fix sizing of sample problem in autotuning
For some choices of tuning parameters the dimensions need to be padded
to a multiple of 3, but that wasn't done with the test problem used for
autotuning. As a result those configurations were skipping during
tuning.

In testing on my laptop, the Prepare kernel now finds one of these
skipped configurations to be best.

See SPR1-230.