PrepareFlags: take visibilities in channel, bl order
This makes it consistent with the ordering used for the flags output.
This required delaying prepare_flags until after transpose_vis.

The Dimensions are also generated on demand, so that the different
arrays won't share Dimension objects (which can cause unnecessary
padding, because padding needed for the one can end up being applied to
the other).