Apply RFI model
The support for getting baseline-length-dependent masks from CAM is
removed since trying to support both the RFI model and these CAM masks
at the same would have been much more complicated. Currently if they are
supplied, just the first mask is used, and a warning is printed.
Eventually katsdpcam2telstate should be updated to switch back to the 1D
format.