Fix some issues in SensorHistory
The main bug was that sensor timestamps were incorrectly converted to
dump indices, because self.rx.timestamp_base was read before it was
initialised (and hence was 0). To avoid such silent bugs in future,
timebase_base now starts life as None rather than 0.

There is a deeper design flaw, in that sensor values may be received
before the first piece of data, at which point it's impossible to
convert timestamps to dump indices. This is fixed by changing the
SensorHistory class to work with timestamps instead of dump indices.
This removes the optimisation that allowed multiple updates of a sensor
during a single dump to be collapsed.

There was a secondary bug that if a sensor sample was received "late"
(i.e. after the value was already queried for the subsequent dump), it
was dropped entirely, instead of added and used for future dumps.