Fix repeated computation by CompressWeights kernel
The C code contains `for` loops to ensure that a single workgroup along
the baseline (fastest-varying) dimension processes the full extent of
the dimension. But the calculation of the global size didn't take this
into account, so it created multiple workgroups along this dimension.
They ended up repeating the same work, possibly hundreds of times,
making compress_weights by far the dominant use of the GPU.

I bumped the autotuning version to force tuning to be re-run, since the
previous tuning results may have been thrown off.